+++ src/cli/main.py
@@ -1121,6 +1121,9 @@ cli.add_command(elb)
 from security.commands import security
 cli.add_command(security)
 
+from audit.commands import audit
+cli.add_command(audit)
+
 # 将 ecs/commands.py 中定义的命令注册到当前 ecs group
 from ecs.commands import (update_ecs_label, query_dedicated_host_uuid, query_order_uuid,
                            cpu_history, mem_history, network_history, disk_history,